added keeper tests
